suitable modularization
