+++ password_manager/cipher.py
@@ -1,6 +1,7 @@
 import threading
 from cryptography.fernet import Fernet
-from .utils import derive_key, reencrypt_passwords, decrypt_master_password
+from .utils import derive_key
+from .master_password_ops import reencrypt_passwords, decrypt_master_password
 class CipherSingleton:
     _instance = None
     _lock = threading.Lock()
@@ -37,3 +38,6 @@ class CipherSingleton:
             self.cipher = self._create_cipher()
             self._cipher_initialized = True
             reencrypt_passwords(old_cipher, self.cipher)
+
+cipher_singleton = CipherSingleton()
+
