+++ password_manager/cli.py
@@ -1,151 +1,21 @@
 import os
-import getpass
 import click
-from cryptography.fernet import Fernet, InvalidToken
-from cryptography.hazmat.primitives.kdf.pbkdf2 import PBKDF2HMAC
-from cryptography.hazmat.primitives import hashes
-from cryptography.hazmat.backends import default_backend
+from cryptography.fernet import InvalidToken
 from pathlib import Path
 import time
 import uuid
-from ctypes import CDLL, c_void_p, c_long
 import threading
 import logging
-import base64
-from .utils import (DATA_DIR, MASTER_PASSWORD_FILE, SESSION_FILE, CURRENT_DIRECTORY_FILE, 
-                    SESSION_TIMEOUT, DATA_DIR, load_secure_key, load_salt, derive_key, set_permissions,
-                    LAContext, LAPolicyDeviceOwnerAuthenticationWithBiometrics, libdispatch, key_lock, 
-                    decrypt_master_password, reencrypt_passwords)
-from .cipher import CipherSingleton
-
-# Instantiate the cipher singleton
-cipher_singleton = CipherSingleton()
-
-def encrypt_master_password(password: str):
-    """Encrypt and store the master password using the secure key."""
-    secure_key = load_secure_key()
-    fernet = Fernet(secure_key)
-    encrypted_password = fernet.encrypt(password.encode())
-    with open(MASTER_PASSWORD_FILE, 'wb') as f:
-        f.write(encrypted_password)
-    set_permissions(MASTER_PASSWORD_FILE)
-    logging.info("Master password stored in encrypted form on disk.")
-
-def load_current_directory():
-    if CURRENT_DIRECTORY_FILE.exists():
-        with open(CURRENT_DIRECTORY_FILE, 'r') as f:
-            return Path(f.read().strip())
-    return DATA_DIR
-
-def save_current_directory(current_directory):
-    with open(CURRENT_DIRECTORY_FILE, 'w') as f:
-        f.write(str(current_directory))
-
-def authenticate_fingerprint_mac():
-    """Authenticate the user using fingerprint on macOS."""
-    context = LAContext.alloc().init()
-    success, error = context.canEvaluatePolicy_error_(LAPolicyDeviceOwnerAuthenticationWithBiometrics, None)
-    
-    if success:
-        click.echo("Please authenticate using your fingerprint...")
-        semaphore = libdispatch.dispatch_semaphore_create(0)
-        
-        def callback(_success, _error):
-            nonlocal authenticated
-            if _success:
-                authenticated = True
-            else:
-                click.echo(f"Fingerprint authentication error: {_error}")
-                authenticated = False
-            libdispatch.dispatch_semaphore_signal(semaphore)
-
-        authenticated = False
-        context.evaluatePolicy_localizedReason_reply_(
-            LAPolicyDeviceOwnerAuthenticationWithBiometrics,
-            "Authenticate to access password manager",
-            callback
-        )
-        libdispatch.dispatch_semaphore_wait(semaphore, c_long(-1))
-        return authenticated
-    else:
-        click.echo(f"Fingerprint authentication not available: {error}")
-        return False
-
-def authenticate_user():
-    try:
-        username = os.getlogin()
-        if username == 'root':
-            username = getpass.getuser()
-    except Exception:
-        username = getpass.getuser()
-
-    choice = click.prompt('Choose authentication method: [P]assword/[F]ingerprint', type=str).lower()
-
-    if choice == 'p':
-        stored_master_password = decrypt_master_password()
-        password = click.prompt('Master Password', hide_input=True)
-        if password != stored_master_password:
-            click.echo("Authentication failed.")
-            exit(1)
-        else:
-            click.echo("Authentication succeeded.")
-            with open(SESSION_FILE, 'w') as f:
-                f.write(str(int(time.time())))
-    elif choice == 'f':
-        if os.uname().sysname == "Darwin":
-            if not authenticate_fingerprint_mac():
-                click.echo("Fingerprint authentication failed.")
-                exit(1)
-            else:
-                click.echo("Fingerprint authentication succeeded.")
-                with open(SESSION_FILE, 'w') as f:
-                    f.write(str(int(time.time())))
-        else:
-            click.echo("Fingerprint authentication is not supported on this system.")
-            exit(1)
-    else:
-        click.echo("Invalid choice.")
-        exit(1)
-
-def ensure_authenticated():
-    """Ensure the user is authenticated before proceeding."""
-    if not is_authenticated():
-        click.echo("You need to authenticate first.")
-        authenticate_user()
-    else:
-        with open(SESSION_FILE, 'w') as f:
-            f.write(str(int(time.time())))
-
-def is_authenticated():
-    if not SESSION_FILE.exists() or os.stat(SESSION_FILE).st_size == 0:
-        return False
-    with open(SESSION_FILE, 'r') as f:
-        timestamp = int(f.read().strip())
-    if time.time() - timestamp > SESSION_TIMEOUT:
-        os.remove(SESSION_FILE)
-        return False
-    return True
-
-
-
-def store_master_password(password: str):
-    """Store the master password and refresh the cipher."""
-    encrypt_master_password(password)
-    cipher_singleton.refresh_cipher()
-
-def get_password_file_path(domain, folder=None):
-    current_dir = load_current_directory()
-    if folder:
-        target_dir = (current_dir / folder).resolve()
-    else:
-        target_dir = current_dir
-    return target_dir / f"{domain}.pass"
+from .utils import (DATA_DIR, MASTER_PASSWORD_FILE, load_secure_key, set_permissions, key_lock, 
+                    load_current_directory, save_current_directory, get_password_file_path)
+from .cipher import cipher_singleton
+from .auth import (ensure_authenticated, authenticate_user)
+from .master_password_ops import store_master_password, reencrypt_passwords
 
 @click.group(invoke_without_command=True)
 @click.pass_context
 def vault(ctx):
     ctx.ensure_object(dict)
-    
     if not MASTER_PASSWORD_FILE.exists():
         click.echo("Master password is not set. Run 'vault set-master-password' to set it.")
         ctx.invoke(set_master_password)
