+++ password_ops.py
@@ -1,0 +1,18 @@
+def encrypt_master_password(password: str):
+    """Encrypt and store the master password using the secure key."""
+    secure_key = load_secure_key()
+    fernet = Fernet(secure_key)
+    encrypted_password = fernet.encrypt(password.encode())
+    with open(MASTER_PASSWORD_FILE, 'wb') as f:
+        f.write(encrypted_password)
+    set_permissions(MASTER_PASSWORD_FILE)
+    logging.info("Master password stored in encrypted form on disk.")
+
+
+def store_master_password(password: str):
+    """Store the master password and refresh the cipher."""
+    encrypt_master_password(password)
+    cipher_singleton.refresh_cipher()
+
+
+
